updated scripts
