+++ scripts/scan_runner.py
@@ -5,10 +5,12 @@ from bluesky import RunEngine
 from dodal.beamlines.p51 import panda, turbo_slit
 from ophyd_async.plan_stubs import ensure_connected
 
-from spectroscopy_bluesky.p51.plans.direct_turbo_slit_movement import (
-    fly_sweep_both_ways,
+from spectroscopy_bluesky.p51.plans.seq_table_scans import (
     seq_non_linear,
     seq_table,
+)
+from spectroscopy_bluesky.p51.plans.turbo_slit_fly_scans import (
+    fly_sweep_both_ways,
     trajectory_fly_scan,
 )
