+++ scripts/test.py
@@ -1,18 +1,14 @@
 from bluesky import RunEngine
-from dodal.beamlines.p51 import panda, turbo_slit
+from dodal.beamlines.p51 import panda, path_provider, turbo_slit
 from ophyd_async.plan_stubs import ensure_connected
 
-from spectroscopy_bluesky.p51.plans.direct_turbo_slit_movement import (
+from spectroscopy_bluesky.p51.plans.turbo_slit_fly_scans import (
     trajectory_fly_scan,
 )
 
 t = turbo_slit()
-p = panda()
+p = panda(path_provider)
 
 RE = RunEngine()
 RE(ensure_connected(t, p))
-RE(
-    trajectory_fly_scan(
-        start=-30, stop=30, num=100, duration=0.01, panda=p, number_of_sweeps=30
-    )
-)
+RE(trajectory_fly_scan(start=-30, stop=30, num=100, duration=0.01, panda=p))
