+++ src/spectroscopy_bluesky/p51/plans/__init__.py
@@ -1,10 +1,16 @@
-from .direct_turbo_slit_movement import (
-    fly_scan_ts,
-    fly_sweep,
-    fly_sweep_both_ways,
+from .common import (
+    plan_store_settings,
+    restore_panda_settings,
+)
+from .seq_table_scans import (
     seq_non_linear,
     seq_table,
     setup_seq_table,
+)
+from .turbo_slit_fly_scans import (
+    fly_scan_ts,
+    fly_sweep,
+    fly_sweep_both_ways,
     trajectory_fly_scan,
 )
 
@@ -16,4 +22,6 @@ __all__ = [
     "seq_table",
     "trajectory_fly_scan",
     "setup_seq_table",
+    "restore_panda_settings",
+    "plan_store_settings",
 ]
